needoh-watcher/requirements.txt
requests>=2.31
beautifulsoup4>=4.12
python-dotenv>=1.0
twilio>=9.0
